MouseLockSystem and Directories changed
