+++ Assets/ZombieHunter/MovementSystem/MovementSystemsContainer.cs
@@ -8,24 +8,13 @@ namespace ZombieHunter.MovementSystem
 {
     public class MovementSystemsContainer : IEcsContainer
     {
-        private readonly EcsStartup.DevelopMode _devMode = null;
-        
         public void AddSystems(EcsSystems ecsSystems)
         {
-            #region MouseLookDevMode
-
-            if (_devMode.Value)
-            {
-                ecsSystems
-                    .Add(new MouseLookSystem.MouseLookSystem())
-                    .Add(new MouseInputSystem());
-            }
-
-            #endregion
-            
             ecsSystems
                 .Add(new BlockJumpSystem())
                 .Add(new PlayerInputSystem.PlayerInputSystem())
+                .Add(new MouseLookSystem.MouseLookSystem())
+                .Add(new CursorLockSystem())
                 .Add(new MovementSystem())
                 .Add(new GroundCheckSystem())
                 .Add(new PlayerJumpSystem());
