--- Assets/ZombieHunter/PlayerInputSystem/MouseInputSystem.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 3ec8d0b8e695420e9ff94bfa14a65365
-timeCreated: 1645546831
